% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_univariate_cates.R
\name{apply_aiptw_transform}
\alias{apply_aiptw_transform}
\title{Apply an AIPTW Transform}
\usage{
apply_aiptw_transform(
  data,
  outcome,
  treatment,
  propensity_score_ls,
  outcome_type
)
}
\arguments{
\item{data}{A \code{tibble} object containing the outcome variable, treatment
indicator, and covariates. The treatment variable is a factor whose levels
correspond to the names of the \code{propensity_score_ls} argument.}

\item{outcome}{A \code{character} defining the name of the outcome variable
in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment levels. Note that the first element
of the list should correspond to the "treatment" group, and the second the
"control" group, whatever their names may be.}

\item{outcome_type}{A \code{character} indicating the type of outcome.
Currently supported outcomes are \code{"continuous"} and
\code{"binomial"}. Here, \code{"binomial"} is used for binary outcomes.}
}
\value{
A \code{tibble} containing the AIPTW transformed predicted outcomes
under treatment and control group assignments of all observations in
\code{data}.
}
\description{
\code{apply_aiptw_transform()} applies an augmented inverse-probability
treatment wright (AIPTW) transform to the estimated potential outcomes
generated by the SuperLearner over the training set.
}
\keyword{internal}
