% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_univariate_survival_cates.R
\name{estimate_univariate_s_cates}
\alias{estimate_univariate_s_cates}
\title{Estimate the Univariate Survival CATEs}
\usage{
estimate_univariate_s_cates(
  long_data,
  event,
  censor,
  treatment,
  biomarkers,
  cond_surv_haz_super_learner,
  cond_censor_haz_super_learner,
  propensity_score_ls,
  v_folds,
  parallel
)
}
\arguments{
\item{long_data}{A \code{tibble} object containing the longitudinal data
output by \code{prep_long_data()}.}

\item{event}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates whether an event occurred.
Observations can have an event or be censored, but not both.}

\item{censor}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a right-censoring event.
Observations can have an event or be censored, but not both.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}.}

\item{cond_surv_haz_super_learner}{A \code{\link[sl3:Lrnr_sl]{Lrnr_sl}}
object used to estimate the conditional event hazard model. If set to
\code{NULL}, the default, an elastic net regression is used instead. It is
best to use this default behaviour when analyzing small datasets.}

\item{cond_censor_haz_super_learner}{A \code{\link[sl3:Lrnr_sl]{Lrnr_sl}}
object used to estimate the conditional censoring hazard model. If set to
\code{NULL}, the default, an elastic net regression is used instead. It is
best to use this default behaviour when analyzing small datasets.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment conditions. The first element of the
list should correspond to the "treatment" condition, and the second to the
"control" condition, whatever their names may be.}

\item{v_folds}{A \code{numeric} indicating the number of folds used for
V-fold cross-validation. Defaults to \code{5}.}

\item{parallel}{A \code{logical} determining whether to use
\code{\link[origami:cross_validate]{origami}}'s built-in parallelized
cross-validation routines. This parallelization framework is built upon
the \href{https://cran.r-project.org/package=future}{\code{future}} suite.
Defaults to \code{FALSE}.}
}
\value{
A \code{list} containing two \code{tibble} objects. The first,
\code{betas_df}, contains the estimated beta coefficients for each
biomarker across all validation sets. The second, \code{ic_df}, is made up
of the cross-validated empirical efficient influence functions of every
observation in \code{data}.
}
\description{
\code{estimate_univariate_survival_cates()} estimates the slope of the
univariate conditional average treatment effect's linear approximation for
right-censored time-to-event outcomes. First, the data is pre-processed
and split into \code{v_folds} folds. Then, the conditional survival and
conditional censoring functions are estimated using their respective
hazard functions' SuperLearners. These estimates are then used to estimate
the expected difference in potential outcomes across treatment conditions
for each individual in their respective validation set. The variable
importance parameters are then estimated in the validation sets, along
with their efficient influence functions.
}
\keyword{internal}
