% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_univariate_survival_cates.R
\name{estimate_univariate_survival_cates}
\alias{estimate_univariate_survival_cates}
\title{Estimate the Univariate Survival CATEs}
\usage{
estimate_univariate_survival_cates(
  long_data,
  failure,
  censor,
  treatment,
  biomarkers,
  cond_surv_haz_super_learner,
  cond_censor_haz_super_learner,
  propensity_score_ls,
  v_folds,
  parallel
)
}
\arguments{
\item{long_data}{A \code{tibble} object containing the longitudinal data
output by \code{prep_long_data()}.}

\item{failure}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a failure event. Observations
can have a failure or a censoring event, but not both.}

\item{censor}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a right-censoring event.
Observations can have a failure or a censoring event, but not both.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}.}

\item{cond_surv_haz_super_learner}{A \code{\link[sl3:Lrnr_sl]{SuperLearner}}
object used to estimate the conditional hazard model. If set to
\code{NULL}, the default SuperLearner is used. The default's library
consists of a linear model, penalized linear models (LASSO and elasticnet),
a Random Forest, and the mean model.}

\item{cond_censor_haz_super_learner}{A
\code{\link[sl3:Lrnr_sl]{SuperLearner}} object used to estimate the
conditional hazard model. If set to \code{NULL}, the default SuperLearner
is used. The default's library consists of a linear model, penalized linear
models (LASSO and elasticnet), a Random Forest, and the mean model.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment levels. The first element of the list
should correspond to the "treatment" group, and the second to the "control"
group, whatever their names may be.}

\item{v_folds}{A \code{numeric} indicating the number of folds to use for
V-fold cross-validation. Defaults to \code{5}.}

\item{parallel}{A \code{logical} determining whether to use
\code{\link[origami:cross_validate]{origami}}'s built-in parallelization
capabilities. Defaults to \code{FALSE}.}
}
\value{
A \code{list} containing two \code{tibbles}. The first,
\code{betas_df}, contains the estimated beta coefficients for each
biomarker across all validation sets. The second, \code{ic_df}, is made up
of the cross-validated influence curves of every observation in
\code{data}.
}
\description{
\code{estimate_univariate_survival cates()} estimates the univariate
Conditional Average Treatment Effect of each biomarker on the predicted
potential time-to-event outcomes using cross-validation. First, the data is
pre-processed and split into \code{v_folds}. Then, the conditional survival
and censoring regressions are estimated over the training sets using
\code{failure_super_learner} and \code{censoring_super_learner},
respectively. The estimated outcome regressions are then used to estimate the
expected difference in potential outcomes across treatment groups for each
individual in their respective validation sets. Univariate linear regressions
are then regressed over the predicted potential outcomes in the validation
set, and the influence curves of all observations is computed as well.
\code{tibbles} of biomarker coefficients and influence curves are returned
for each validation set.
}
\keyword{internal}
