% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_univariate_cates.R
\name{hold_out_calculation}
\alias{hold_out_calculation}
\title{Compute Validation Set Objects}
\usage{
hold_out_calculation(
  fold,
  data,
  outcome,
  treatment,
  biomarkers,
  super_learner,
  propensity_score_ls,
  outcome_type
)
}
\arguments{
\item{fold}{A \code{fold} object (from \code{\link[origami]{make_folds}()}).}

\item{data}{A \code{tibble} object containing the outcome variable, treatment
indicator, and covariates. The treatment variable is a factor whose levels
correspond to the names of the \code{propensity_score_ls} argument.}

\item{outcome}{A \code{character} defining the name of the outcome variable
in \code{data}. The outcome must be a continuous or a binary factor
variable.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the pre-treatment
biomarkers variables in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{super_learner}{A \code{\link[sl3:Lrnr_sl]{Lrnr_sl}} object used to
estimate the conditional outcome regression. If set to \code{NULL}, the
default, an elastic net regression is used instead. It is best to use this
default behaviour when analyzing small datasets.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment conditions. The first element of the
list should correspond to the "treatment" condition, and the second to the
"control" condition, whatever their names may be.}

\item{outcome_type}{A \code{character} indicating the type of outcome.
Currently supported outcomes are \code{"continuous"} and
\code{"binomial"}. Here, \code{"binomial"} is used for binary outcomes.}
}
\value{
A \code{list} made up of two objects. The first is the \code{numeric}
vector of biomarker variable importance estimates. The second is the
\code{tibble} of the empirical efficient influence functions for each
biomarker.
}
\description{
\code{hold_out_calculation()} computes all the objects required to estimate
the variable importance parameter and compute their empirical efficient
influence curve over the validation set, assuming continuous or binary
outcomes. It begins by estimated the conditional outcome regression on the
training set using the \code{super_learner}. Next, the difference in
potential outcomes is predicted for the observations in the validation set.
Finally, these predicted outcomes are used to estimate the variable
importance parameters and efficient influence functions.
}
\keyword{internal}
