% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_inference.R
\name{perform_inference}
\alias{perform_inference}
\title{Perform Inference on Estimated Biomarker Coefficients}
\usage{
perform_inference(biomarkers_tbl)
}
\arguments{
\item{biomarkers_tbl}{A \code{tibble} of estimated coefficients and their
standard errors.}
}
\value{
A \code{tibble} of estimated coefficients, standard errors, test
statistics, and (adjusted) p-values for each biomarker. The biomarkers
are ordered by significance.
}
\description{
\code{perform_inference()} performs valid inference using the estimated
biomarker coefficients and the standard errors computed using the
cross-validated influence curves. The computed test statistics are
asymptotically Normal with mean zero under the null hypothesis. The resulting
p-values for the two-sided tests are subsequently corrected for multiple
testing using the Benjamini-Hochberg method to control the False Discovery
Rate, and Holm's procedure to compute control the Family-Wise Error Rate.
}
\keyword{internal}
