% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_long_data.R
\name{prep_long_data}
\alias{prep_long_data}
\title{Prepare Longitudinal Time-to-Event Data}
\usage{
prep_long_data(
  data,
  status,
  relative_time,
  treatment,
  covariates,
  biomarkers,
  time_cutoff = NULL
)
}
\arguments{
\item{data}{A wide \code{data.frame} or \code{tibble} object containing the
status (event variable), relative time of the event, treatment indicator,
and covariates. Note that the biomarkers must be a subset of the
covariates, and that there should only be one row per observation.}

\item{status}{A \code{character} defining the name of the status variable
in \code{data}. This binary variable indicates whether the observation
failed at the associated \code{relative_time}, or if it was censored.
Failures should be represented by a \code{1}, and censoring events by a
\code{0}.}

\item{relative_time}{A \code{character} providing the name of the time
variable in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{covariates}{A \code{character} vector listing the covariates in
\code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{time_cutoff}{A \code{numeric} representing the time at which to assess
the biomarkers' importance with respect to the outcome. If not specified,
this value is set to the maximum value in the \code{data} argument's
\code{relative_time} variable.}
}
\value{
A longitudinal \code{tibble} containing only the status variable,
a relative time variable, treatment indicator, covariates, and observation
identifier. The treatment variable is transformed into a factor if not
already, and no rows whose relative time values are larger than
\code{time_cutoff} are retained when \code{time_cutoff} is non-null.
}
\description{
Prepare Longitudinal Time-to-Event Data
}
\keyword{internal}
