% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunicate.R
\name{sunicate}
\alias{sunicate}
\title{Univariate Conditional Average Treatment Effect Estimation For Right-Censored Time-To-Event Outcomes}
\usage{
sunicate(
  data,
  failure,
  censor,
  relative_time,
  treatment,
  covariates,
  biomarkers,
  time_cutoff = NULL,
  cond_surv_haz_super_learner = NULL,
  cond_censor_haz_super_learner = NULL,
  propensity_score_ls,
  v_folds = 5L,
  parallel = FALSE
)
}
\arguments{
\item{data}{A wide \code{data.frame} or \code{tibble} object containing the
status (event variable), relative time of the event, treatment indicator,
and covariates. Note that the biomarkers must be a subset of the
covariates, and that there should only be one row per observation.}

\item{failure}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a failure event. Observations
can have a failure or a censoring event, but not both.}

\item{censor}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a right-censoring event.
Observations can have a failure or a censoring event, but not both.}

\item{relative_time}{A \code{character} providing the name of the time
variable in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{covariates}{A \code{character} vector listing the covariates in
\code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{time_cutoff}{A \code{numeric} representing the time at which to assess
the biomarkers' importance with respect to the outcome. If not specified,
this value is set to the maximum value in the \code{data} argument's
\code{relative_time} variable.}

\item{cond_surv_haz_super_learner}{A \code{\link[sl3:Lrnr_sl]{SuperLearner}}
object used to estimate the conditional hazard model. If set to
\code{NULL}, the default SuperLearner is used. The default's library
consists of a linear model, penalized linear models (LASSO and elasticnet),
a Random Forest, and the mean model.}

\item{cond_censor_haz_super_learner}{A
\code{\link[sl3:Lrnr_sl]{SuperLearner}} object used to estimate the
conditional hazard model. If set to \code{NULL}, the default SuperLearner
is used. The default's library consists of a linear model, penalized linear
models (LASSO and elasticnet), a Random Forest, and the mean model.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment levels. The first element of the list
should correspond to the "treatment" group, and the second to the "control"
group, whatever their names may be.}

\item{v_folds}{A \code{numeric} indicating the number of folds to use for
V-fold cross-validation. Defaults to \code{5}.}

\item{parallel}{A \code{logical} determining whether to use
\code{\link[origami:cross_validate]{origami}}'s built-in parallelization
capabilities. Defaults to \code{FALSE}.}
}
\value{
A \code{tibble} of the biomarkers. Each row contains the estimated
univariate linear model coefficient of the biomarker regressed against the
predicted difference in adjsuted survival probabilities, along with the
standard error, test statistic, and (adjusted) p-values. The table is
ordered by significance.
}
\description{
\code{sunicate()} estimates a linear approximation of the conditional
average treatment effect of a time-to-event outcome for each variable
specified in the \code{biomarkers} argument. The estimated coefficients are
then used to test whether their corresponding biomarkers are
treatment-effect modifiers. Valid statistical inference is performed by
employing the biomarkers' cross-validated empirical influence curves to
estimate these coefficients' standard errors.
}
