% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coefs_table.R
\name{compute_coefs_tbl}
\alias{compute_coefs_tbl}
\title{Compute Linear Model Coefficients and Standard Errors}
\usage{
compute_coefs_tbl(cv_ls)
}
\arguments{
\item{cv_ls}{A \code{list} containing two \code{tibbles}. The first,
\code{betas_df}, contains the estimated beta coefficients for each
biomarker across all validation sets. The second, \code{ic_df}, is made up
of the cross-validated influence curves of every observation in
\code{data}.}
}
\value{
A \code{tibble} of estimated coefficients and their standard errors.
}
\description{
\code{compute_coefs_tbl()} estimates the univariate linear model coefficients
of each biomarker, and computes their accompanying standard errors.
}
\keyword{internal}
