% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_univariate_cates.R
\name{estimate_univariate_cates}
\alias{estimate_univariate_cates}
\title{Estimate the Univariate CATEs}
\usage{
estimate_univariate_cates(
  data,
  outcome,
  treatment,
  biomarkers,
  super_learner,
  propensity_score_ls,
  v_folds,
  parallel
)
}
\arguments{
\item{data}{A \code{tibble} object containing the outcome variable, treatment
indicator, and covariates. The treatment variable is a factor whose levels
correspond to the names of the \code{propensity_score_ls} argument.}

\item{outcome}{A \code{character} defining the name of the outcome variable
in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}.}

\item{super_learner}{A \code{\link[sl3:Lrnr_sl]{SuperLearner}} object. If set
to \code{NULL}, the default SuperLearner is used.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment levels. Note that the first element
of the list should correspond to the "treatment" group, and the second the
"control" group, whatever their names may be.}

\item{v_folds}{A \code{numeric} indicating the number of folds to use for
V-fold cross-validation.}

\item{parallel}{A \code{logical} determining whether to use
\code{\link[origami:cross_validate]{origami}}'s built-in parallelization
capabilities.}
}
\value{
A \code{list} containing two \code{tibbles}. The first,
\code{betas_df}, contains the estimated beta coefficients for each
biomarker across all validation sets. The second, \code{ic_df}, is made up
of the cross-validated influence curves of every observation in
\code{data}.
}
\description{
\code{estimate_univariate_cates()} estimates the univariate Conditional
Average Treatment Effect of each biomarker on the predicted potential
outcomes using cross-validation. First, the data is pre-processed data is
split into \code{v_folds}. Then, the conditional outcome regression over all
the training sets are estimated using the \code{super_learner}. The estimated
outcome regressions are then used to estimate the expected difference in
potential outcomes across treatment groups for each individual in their
respective validation sets. Univariate linear regressions are then regressed
over the predicted potential outcomes in the validation set, and the
influence curves of all observations is computed as well. \code{tibbles} of
biomarker coefficients and influence curves are returned for each validation
set.
}
\keyword{internal}
