% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_univariate_cates.R
\name{hold_out_calculation}
\alias{hold_out_calculation}
\title{Compute Validation Set Objects}
\usage{
hold_out_calculation(
  fold,
  data,
  outcome,
  treatment,
  biomarkers,
  super_learner,
  propensity_score_ls,
  outcome_type
)
}
\arguments{
\item{fold}{A \code{fold} object (from \code{\link[origami]{make_folds}()}).}

\item{data}{A \code{tibble} object containing the outcome variable, treatment
indicator, and covariates. The treatment variable is a factor whose levels
correspond to the names of the \code{propensity_score_ls} argument.}

\item{outcome}{A \code{character} defining the name of the outcome variable
in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}.}

\item{super_learner}{A \code{\link[sl3:Lrnr_sl]{SuperLearner}} object. If set
to \code{NULL}, the default SuperLearner is used.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment levels. Note that the first element
of the list should correspond to the "treatment" group, and the second the
"control" group, whatever their names may be.}

\item{outcome_type}{A \code{character} indicating the type of outcome.
Currently supported outcomes are \code{"continuous"} and
\code{"binomial"}. Here, \code{"binomial"} is used for binary outcomes.}
}
\value{
A \code{list} made up of two objects. The first is the \code{numeric}
vector of biomarker linear model coefficient estimates. The second is the
\code{tibble} of the empirical influence curves for each biomarker.
}
\description{
\code{hold_out_calculation} computes all the objects required to estimated
the univariate CATE over the validation set. It begins by estimated the
conditional outcome regression on the training set using the
\code{super_learner}. Next, the difference in potential outcomes is
predicted for the observations in the validation set. Finally, these
predicted outcomes are used to estimate the unvariate regression coefficients
of each biomarker, and their respective empirical influence curves are
computed as well.
}
\keyword{internal}
