% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare Data}
\usage{
prep_data(
  data,
  outcome,
  treatment,
  covariates,
  biomarkers,
  propensity_score_ls = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} object containing the
outcome variable, treatment indicator, and pre-treatment covariates. Note
that the biomarkers are considered be a subset of the covariates.}

\item{outcome}{A \code{character} defining the name of the outcome variable
in \code{data}. The outcome must be a continuous or a binary factor
variable.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{covariates}{A \code{character} vector listing the pre-treatment
covariates variables in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the pre-treatment
biomarkers variables in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment conditions. The first element of the
list should correspond to the "treatment" condition, and the second to the
"control" condition, whatever their names may be.}
}
\value{
A \code{tibble} containing only the outcome variable, treatment
indicator, and covariates. Note that the treatment variable is transformed
into a factor if not already.
}
\description{
\code{prep_data()} assesses whether the data passed to \code{\link{unicate}}
is correctly formatted. This function also performs minor modifications to
the data, (e.g. transforming the treatment variable into a factor if not
already).
}
\keyword{internal}
