% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_long_data.R
\name{prep_long_data}
\alias{prep_long_data}
\title{Prepare Longitudinal Time-to-Event Data}
\usage{
prep_long_data(
  data,
  event,
  censor,
  relative_time,
  treatment,
  covariates,
  biomarkers,
  time_cutoff = NULL,
  propensity_score_ls = NULL
)
}
\arguments{
\item{data}{A wide \code{data.frame} or \code{tibble} object containing the
status (event variable), relative time of the event, treatment indicator,
and covariates. Note that the biomarkers must be a subset of the
covariates, and that there should only be one row per observation.}

\item{event}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates whether an event occurred.
Observations can have an event or be censored, but not both.}

\item{censor}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a right-censoring event.
Observations can have an event or be censored, but not both.}

\item{relative_time}{A \code{character} providing the name of the time
variable in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{covariates}{A \code{character} vector listing the covariates in
\code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{time_cutoff}{A \code{numeric} representing the time at which to assess
the biomarkers' importance with respect to the outcome. If not specified,
this value is set to the median value in the \code{data} argument's
\code{relative_time} variable.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment conditions. The first element of the
list should correspond to the "treatment" condition, and the second to the
"control" condition, whatever their names may be.}
}
\value{
A longitudinal \code{tibble} containing only the event indicator
variables, a relative time variable, treatment indicator, covariates, and
observation identifier. The treatment variable is transformed into a factor
if not already, and no rows whose relative time values are larger than
\code{time_cutoff} are retained when \code{time_cutoff} is non-null.
}
\description{
\code{prep_long_data()} assesses whether the data passed to
\code{\link{sunicate}} is correctly formatted. This function also performs
modifications to the data, (e.g. transforming the treatment variable into a
factor if not already, and transforming it from a wide to a long format).
}
\keyword{internal}
