% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunicate.R
\name{sunicate}
\alias{sunicate}
\title{Univariate Conditional Average Treatment Effect Estimation For Right-Censored
Time-To-Event Outcomes}
\usage{
sunicate(
  data,
  event,
  censor,
  relative_time,
  treatment,
  covariates,
  biomarkers,
  time_cutoff = NULL,
  cond_surv_haz_super_learner = NULL,
  cond_censor_haz_super_learner = NULL,
  propensity_score_ls,
  v_folds = 5L,
  parallel = FALSE
)
}
\arguments{
\item{data}{A "wide" \code{data.frame} or \code{tibble} object containing the
status (event variable), relative time of the event, treatment indicator,
and covariates. Note that the biomarkers must be a subset of the
covariates, and that there should only be one row per observation.}

\item{event}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates whether an event occurred.
Observations can have an event or be censored, but not both.}

\item{censor}{A \code{character} defining the name of the binary variable in
the \code{data} argument that indicates a right-censoring event.
Observations can have an event or be censored, but not both.}

\item{relative_time}{A \code{character} providing the name of the time
variable in \code{data}.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{covariates}{A \code{character} vector listing the covariates in
\code{data}.}

\item{biomarkers}{A \code{character} vector listing the biomarkers of
interest in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{time_cutoff}{A \code{numeric} representing the time at which to assess
the biomarkers' importance with respect to the outcome. If not specified,
this value is set to the median value of the \code{data} argument's
\code{relative_time} variable.}

\item{cond_surv_haz_super_learner}{A \code{\link[sl3:Lrnr_sl]{Lrnr_sl}}
object used to estimate the conditional event hazard model. If set to
\code{NULL}, the default SuperLearner is used. The default's library
consists of a linear model, penalized linear models (LASSO and elasticnet),
a Random Forest, and the mean model.}

\item{cond_censor_haz_super_learner}{A \code{\link[sl3:Lrnr_sl]{Lrnr_sl}}
object used to estimate the conditional censoring hazard model. If set to
\code{NULL}, the default SuperLearner is used. The default's library
consists of a linear model, penalized linear models (LASSO and elasticnet),
a Random Forest, and the mean model.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment conditions. The first element of the
list should correspond to the "treatment" condition, and the second to the
"control" condition, whatever their names may be.}

\item{v_folds}{A \code{numeric} indicating the number of folds used for
V-fold cross-validation. Defaults to \code{5}.}

\item{parallel}{A \code{logical} determining whether to use
\code{\link[origami:cross_validate]{origami}}'s built-in parallelized
cross-validation routines. This parallelization framework is built upon
the \href{https://cran.r-project.org/package=future}{\code{future}} suite.
Defaults to \code{FALSE}.}
}
\value{
A \code{tibble} with rows corresponding to the specified
\code{biomarkers}. Each row contains an estimate of the
treatment-modification variable importance parameter, its standard error,
z-score, and the nominal and adjusted p-values of the accompanying test.
FDR and FWER adjustments are performed using the Benjamini-Hochberg method
and Holm's procedure, respectively. The \code{biomarkers} are ordered by
significance.
}
\description{
\code{sunicate()} performs inference about \code{biomarkers}' capacity to
modify treatment effects in randomized control trials with right-censored,
time to event outcomes. The strength of each \code{biomarkers}' treatment
effect modification is captured by an unknown variable importance parameter
defined as the slope of the univariate conditional average treatment
effect's linear approximation. In all but pathological cases, the larger
the absolute value of this parameter, the greater the treatment effect
modification. \code{unicate()} implements assumption-lean, cross-validated
inference procedures about these variable importance parameters based on
semiparametric theory. Assuming that the \code{biomarkers} have non-zero
variance and that either the conditional survival or conditional censoring
functions are well-estimated by their respective SuperLearners, estimates
are generated by an unbiased and consistent estimator. Tests assessing
whether these slope parameters are significantly different from zero are
also performed; they are valid under the same bounded-variance condition
and assuming that the censoring mechanism is consistently estimated.
}
