% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicate.R
\name{unicate}
\alias{unicate}
\title{Univariate Conditional Average Treatment Effect Estimation}
\usage{
unicate(
  data,
  outcome,
  treatment,
  covariates,
  biomarkers,
  super_learner = NULL,
  propensity_score_ls,
  v_folds = 5L,
  parallel = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} object containing the
outcome variable, treatment indicator, and pre-treatment covariates. Note
that the biomarkers are considered be a subset of the covariates.}

\item{outcome}{A \code{character} defining the name of the outcome variable
in \code{data}. The outcome must be a continuous or a binary factor
variable.}

\item{treatment}{A \code{character} indicating the name of the binary
treatment variable in \code{data}.}

\item{covariates}{A \code{character} vector listing the pre-treatment
covariates variables in \code{data}.}

\item{biomarkers}{A \code{character} vector listing the pre-treatment
biomarkers variables in \code{data}. \code{biomarkers} must be a subset of
\code{covariates}.}

\item{super_learner}{A \code{\link[sl3:Lrnr_sl]{Lrnr_sl}} object used to
estimate the conditional outcome regression. If set to \code{NULL}, the
default, an elastic net regression is used instead. It is best to use this
default behaviour when analyzing small datasets.}

\item{propensity_score_ls}{A named \code{numeric} \code{list} providing the
propensity scores for the treatment conditions. The first element of the
list should correspond to the "treatment" condition, and the second to the
"control" condition, whatever their names may be.}

\item{v_folds}{A \code{numeric} indicating the number of folds used for
V-fold cross-validation. Defaults to \code{5}.}

\item{parallel}{A \code{logical} determining whether to use
\code{\link[origami:cross_validate]{origami}}'s built-in parallelized
cross-validation routines. This parallelization framework is built upon
the \href{https://cran.r-project.org/package=future}{\code{future}} suite.
Defaults to \code{FALSE}.}
}
\value{
A \code{tibble} with rows corresponding to the specified
\code{biomarkers}. Each row contains an estimate of the
treatment-modification variable importance parameter, its standard error,
z-score, and the nominal and adjusted p-values of the accompanying test.
FDR and FWER adjustments are performed using the Benjamini-Hochberg method
and Holm's procedure, respectively. The \code{biomarkers} are ordered by
significance.
}
\description{
\code{unicate()} performs inference about \code{biomarkers}' capacity to
modify treatment effects in randomized control trials with continuous or
binary outcomes. The strength of each \code{biomarkers}' treatment effect
modification is captured by an unknown variable importance parameter
defined as the slope of the univariate conditional average treatment
effect's linear approximation. In all but pathological cases, the larger
the absolute value of this parameter, the greater the treatment effect
modification. \code{unicate()} implements assumption-lean, cross-validated
inference procedures about these variable importance parameters based on
semiparametric theory. Assuming that the \code{biomarkers} have non-zero
variance, estimates are generated by an unbiased and consistent estimator.
Tests assessing whether these slope parameters are significantly different
from zero are also performed; they are valid under the same
bounded-variance condition.
}
